Remodel the client to be plugable
