+++ lib/manageiq/messaging.rb
@@ -1,6 +1,5 @@
 require 'active_support/core_ext/module/delegation'
 require 'yaml'
-require 'stomp'
 
 require 'manageiq/messaging/null_logger'
 
@@ -17,9 +16,4 @@ module ManageIQ
 end
 
 require 'manageiq/messaging/version'
-require 'manageiq/messaging/common'
 require 'manageiq/messaging/client'
-require 'manageiq/messaging/stomp_client'
-require 'manageiq/messaging/queue'
-require 'manageiq/messaging/topic'
-require 'manageiq/messaging/background_job'
