+++ lib/manageiq/messaging/client.rb
@@ -1,23 +1,29 @@
 module ManageIQ
   module Messaging
     class Client
+      require 'manageiq/messaging/stomp/client'
+
       # Open or create a connection to the message broker
+      # @param type [String or Symbol] client type, available choices are:
+      #   :Stomp
+      #   :AMQP
+      #   :MiqQueue
       # @param options [Hash] the connection options
       # @return [Client, nil] the client object if no block is given
       #   The optional block supply {|client| block }. The client will
       #   be automatically closed when the block terminates
-      def self.open(options)
-        # TODO: replace with a AMQP client
-        client = StompClient.new(options)
+      #
+      # Avaiable type:
+      def self.open(type = :Stomp, options)
+        client = Object.const_get("ManageIQ::Messaging::#{type}::Client").new(options)
         return client unless block_given?
-        if block_given?
-          begin
-            yield client
-          ensure
-            client.close
-          end
-          nil
+
+        begin
+          yield client
+        ensure
+          client.close
         end
+        nil
       end
 
       # Publish to a message to a queue. The message will be delivered to only one
@@ -37,7 +43,9 @@ module ManageIQ
       # Optionally a call back block {|response| block} can be provided to wait on
       # the consumer to send an acknowledgment.
       def publish_message(options, &block)
-        Queue.publish(self, options, &block)
+        assert_options(options, [:message, :service])
+
+        publish_message_impl(options, &block)
       end
 
       # Publish multiple messages to a queue.
@@ -46,7 +54,7 @@ module ManageIQ
       #
       # @param messages [Array] a collection of options for `#publish_message`
       def publish_messages(messages)
-        Queue.publish_batch(self, messages)
+        publish_messages_impl(messages)
       end
 
       # Subscribe to receive messages from a queue
@@ -60,16 +68,20 @@ module ManageIQ
       # A callback block {|messages| block} needs to be provided to consume the
       # messages. Example
       #   subscribe_message(options) do |messages|
-      #     messages.each do |msg|
+      #     messages.collect do |msg|
       #       # from msg you get
       #       msg.sender
       #       msg.message
       #       msg.payload
       #       msg.ack_ref (used to ack the message)
+      #
+      #       client.ack(msg.ack_ref)
+      #       # process
+      #       result # a result sent back to sender if expected
       #     end
       #   end
       #
-      # @note The subscriber should ack each message independently in the callback
+      # @note The subscriber MUST ack each message independently in the callback
       # block. It can decide when to ack according to whether a message can
       # be retried. Ack the message in the beginning of processing if the
       # message is not re-triable; otherwise ack it after the message is done.
@@ -80,8 +92,9 @@ module ManageIQ
       # To ack a message call `ack(msg.ack_ref)`
       def subscribe_messages(options, &block)
         raise "A block is required" unless block_given?
+        assert_options(options, [:service])
 
-        Queue.subscribe(self, options, &block)
+        subscribe_messages_impl(options, &block)
       end
 
       # Subscribe to receive from a queue and run each message as a background job.
@@ -94,7 +107,9 @@ module ManageIQ
       # Background job assumes each job is not re-triable. It will ack as soon as a request
       # is received
       def subscribe_background_job(options)
-        BackgroundJob.subscribe(self, options)
+        assert_options(options, [:service])
+
+        subscribe_background_job_impl(options)
       end
 
       # Publish a message as a topic. All subscribers will receive a copy of the message.
@@ -106,7 +121,9 @@ module ManageIQ
       #   <other queue options TBA>
       #
       def publish_topic(options)
-        Topic.publish(self, options)
+        assert_options(options, [:event, :service])
+
+        publish_topic_impl(options)
       end
 
       # Subscribe to receive topic type messages.
@@ -121,8 +138,23 @@ module ManageIQ
       #
       def subscribe_topic(options, &block)
         raise "A block is required" unless block_given?
+        assert_options(options, [:service])
 
-        Topic.subscribe(self, options, &block)
+        subscribe_topic_impl(options, &block)
+      end
+
+      Struct.new("ManageIQ_Messaging_ReceivedMessage", :sender, :message, :payload, :ack_ref)
+
+      private
+
+      def logger
+        ManageIQ::Messaging.logger
+      end
+
+      def assert_options(options, keys)
+        keys.each do |key|
+          raise "options must contain key #{key}" unless options.key?(key)
+        end
       end
     end
   end
